requirements.txt
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
debugpy==1.6.5
dill==0.3.6
git-filter-repo==2.38.0
gitdb==4.0.10
GitPython==3.1.30
prompt-toolkit==3.0.36
Pygments==2.14.0
questionary==1.10.0
rich==13.0.1
smmap==5.0.0
wcwidth==0.2.5
